+++ packages/pdf-parser/src/model/ParseContext.js
@@ -5,6 +5,7 @@ class Context {
     this.options = undefined;
     this.entries = undefined;
     this.document = undefined;
+    this.worker = undefined;
     const rawBook = {};
     Object.keys(new Book()).forEach((key) => { rawBook[key] = undefined; });
     this.rawBook = rawBook;